luwu/pyproject.toml
[project]
name = "luwu"
version = "0.1.0"
description = "Advanced Legged Robot Parkour Training with Isaac Sim and Isaac Lab"
authors = [
    { name = "ZheyangXu", email = "jishengxzy@hotmail.com" },
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "dynaconf>=3.2.0",
    "pyyaml>=6.0.0",
    "wandb>=0.16.0",
    "tensorboard>=2.14.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.0",
    "opencv-python>=4.8.0",
    "scipy>=1.11.0",
    "tqdm>=4.66.0",
    "gymnasium>=0.29.0",
    "mujoco>=3.0.0",
    # Isaac Sim/Lab dependencies - commented out for demo
    # "isaacsim",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "black>=23.9.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
luwu-train = "luwu.interfaces.cli:train"
luwu-play  = "luwu.interfaces.cli:play"
luwu-eval  = "luwu.interfaces.cli:evaluate"

[build-system]
requires      = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.black]
line-length    = 100
target-version = ["py312"]

[tool.isort]
atomic                 = true
profile                = "black"
src_paths              = ["src", "tests"]
extra_standard_library = ["typing_extensions"]
indent                 = 4
lines_after_imports    = 2
multi_line_output      = 3

[tool.ruff]
lint.select = ["C", "E", "F", "I", "W"]
lint.ignore = ["C901", "E501", "E741", "F402", "F823"]
lint.isort  = { lines-after-imports = 2 }

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style               = "double"
indent-style              = "space"
skip-magic-trailing-comma = false
line-ending               = "auto"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths  = ["tests"]
addopts    = "--cov=luwu --cov-report=term-missing --cov-report=html --cov-fail-under=90"

[tool.coverage.run]
source = ["src/luwu"]
omit = [
    "tests/*",
    "src/luwu/__main__.py",
]

[tool.mypy]
python_version              = "3.10"
warn_return_any             = true
warn_unused_configs         = true
disallow_untyped_defs       = true
disallow_incomplete_defs    = true
check_untyped_defs          = true
disallow_untyped_decorators = true
no_implicit_optional        = true
warn_redundant_casts        = true
warn_unused_ignores         = true
warn_no_return              = true
warn_unreachable            = true
strict_equality             = true

[tool.pdm.scripts]
lint     = "ruff check src tests"
test     = "pytest"
format   = "black src tests"
isort    = "isort src tests"
mypy     = "mypy src"
coverage = "coverage run --source=src -m pytest -v tests"
all      = { composite = ["format", "isort", "lint", "mypy", "test"] }
